+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to the "og-devops-status" extension will be documented in this file.
 
+## [0.3.0]
+
+- Releases now also shows owner
+- Status now includes project name besides build/release definition name
+
 ## [0.2.0]
 
 - Differentiate between running builds and releases